Use bridge network for all containers
This is important to share the same network between containers to do inter-cluster routing properly. If networks are not specified, each docker-compose file results in a separate bridge network (this is why I guess we used "host" network).

Also, I've used the hostnames instead of localhost in multiple places since it's a prerequsite for moving to the cloud.

Testing:
* ran server/frontend/database/worker/message-broker